requirements-dev.txt
django-stubs==4.2.7
Faker==25.2.0
